+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 1.2.0 - 2023-05-07
+
+### Added
+
+- Added ability to register identifiers.
+- Added INSTALL, MAPPER\_ADAPTOR, SOURCE\_MAP, and SNYPPET identifiers.
+- Added snyppet and install middlewares.
+
+### Changed
+
+- Removed unused PAGE identifier.
+- Added default values to some existing middleware constructors.
+
 ## 1.1.0 - 2023-01-15
 
 ### Added